需要重写设备号的驱动，区分主设备与次设备，所有主设备

+++ lib/dietlibc/include/sys/types.h
@@ -50,11 +50,16 @@ typedef unsigned long int clock_t;	/* Used for system times in
     typedef uint32_t nlink_t;		/* Used for link counts. */
     typedef uint32_t uid_t;		/* Used for user IDs. */
 #elif defined(__arm__) || defined(__i386__) || defined(__sparc__) || defined(__s390__) /* make sure __s390x__ hits before __s390__ */
-    typedef uint16_t dev_t;
-    typedef uint16_t gid_t;
-    typedef uint16_t mode_t;
-    typedef uint16_t nlink_t;
-    typedef uint16_t uid_t;
+//    typedef uint16_t dev_t;
+//    typedef uint16_t gid_t;
+//    typedef uint16_t mode_t;
+//    typedef uint16_t nlink_t;
+//    typedef uint16_t uid_t;zhangzheng
+    typedef uint32_t dev_t;
+    typedef uint32_t gid_t;
+    typedef uint32_t mode_t;
+    typedef uint32_t nlink_t;
+    typedef uint32_t uid_t;
 #elif defined(__hppa__)
     typedef uint32_t dev_t;
     typedef uint32_t gid_t;
